Move Retrofit Classes to AamnapmRetrofit Module
