+++ AamnapmRetrofit/src/main/java/com/aamnapm/aamnapmretrofit/Constant.java
@@ -1,4 +1,4 @@
-package ir.aamnapm.retrofitsamples.di;
+package com.aamnapm.aamnapmretrofit;
 
 public class Constant {
     public static String BASE_URL = "https://jsonplaceholder.typicode.com/";
