+++ AamnapmRetrofit/src/main/java/com/aamnapm/aamnapmretrofit/CustomCall.java
@@ -1,0 +1,13 @@
+package com.aamnapm.aamnapmretrofit;
+
+import androidx.annotation.NonNull;
+
+public interface CustomCall<T> {
+
+    void cancel();
+
+    void enqueue(CustomCallback<T> callback);
+
+    @NonNull
+    CustomCall<T> clone();
+}
