+++ AamnapmRetrofit/src/main/java/com/aamnapm/aamnapmretrofit/RetrofitInstance.java
@@ -1,16 +1,14 @@
-package ir.aamnapm.retrofitsamples.di.module;
+package com.aamnapm.aamnapmretrofit;
 
-import static ir.aamnapm.retrofitsamples.di.Constant.BASE_URL;
 
-import ir.aamnapm.retrofitsamples.rerofitUtils.ErrorHandlingCallAdapterFactory;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
-public class AppModule {
+public class RetrofitInstance {
 
     public static Retrofit provideRetrofitInstance() {
         return new Retrofit.Builder()
-                .baseUrl(BASE_URL)
+                .baseUrl(Constant.BASE_URL)
                 .addCallAdapterFactory(new ErrorHandlingCallAdapterFactory())
                 .addConverterFactory(GsonConverterFactory.create())
                 .build();
