+++ AamnapmRetrofit/src/main/java/com/aamnapm/aamnapmretrofit/ViewModelProviderFactory.java
@@ -1,4 +1,4 @@
-package ir.aamnapm.retrofitsamples.di.config;
+package com.aamnapm.aamnapmretrofit;
 
 
 import androidx.annotation.NonNull;
