--- app/src/main/java/ir/aamnapm/retrofitsamples/rerofitUtils/MyCall.java
@@ -1,13 +1,0 @@
-package ir.aamnapm.retrofitsamples.rerofitUtils;
-
-import androidx.annotation.NonNull;
-
-public interface MyCall<T> {
-
-    void cancel();
-
-    void enqueue(MyCallback<T> callback);
-
-    @NonNull
-    MyCall<T> clone();
-}
